FEAT : #244 - 상세보기 파일 첨부 출력
